refactor(mso_mdoc): split key_routes.py into key and trust anchor modules
key_routes.py mixed signing key/certificate management with trust anchor
management — two unrelated concerns in one 460-line file.

- key_routes.py: now covers only signing key and certificate schemas,
  handlers, and route registration (register_key_routes).
- trust_anchor_routes.py (new): trust anchor schemas, handlers, and
  route registration (register_trust_anchor_routes).
- routes.py: imports and calls both register functions separately.
- register_key_management_routes alias kept in key_routes.py for
  backward compatibility.
